apps/backend/requirements.txt
# Core Framework
fastapi==0.120.4
uvicorn[standard]==0.38.0
sse-starlette==2.2.0

# External Services
supabase==2.23.0
google-generativeai==0.8.5

# Configuration & Environment
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Utilities
python-multipart==0.0.18
# Pin pydantic to satisfy supabase->realtime (requires >=2.11.7)
pydantic[email]==2.11.7
# authlib 1.3.x does not expose a 'jose' extra; base package includes JOSE utilities
authlib==1.3.2
httpx==0.28.1

# SVG Sanitization
bleach==6.2.0
# tinycss2 bumped to meet weasyprint>=62.3 requirement (needs >=1.3.0)
tinycss2==1.3.0

# PDF Generation (Sprint 6)
weasyprint==62.3
markdown==3.7

# Text Extraction Libraries
pymupdf==1.24.14
python-docx==1.1.2
pytesseract==0.3.13
pillow==11.0.0

# Text Processing
tiktoken==0.8.0

# Testing (Sprint 2+)
pytest==8.3.4
pytest-asyncio==0.24.0

# Development Tools (to be added in Sprint 7)
# ruff==X.Y.Z
# mypy==X.Y.Z
# black==X.Y.Z
